+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 0.1.0
+# 0.1.2 10/24/23
+- Added applied screen
+- fixed .gitignore bug where whitespace was counted
+
+# 0.1.1 10/23/23
+- Added back cursor
+
+# 0.1.0 10/23/23
 - First release of Headifier.
 - core functionality works and is tested
 - there might be bugs with rendering, let me know!